requirements.txt
# Google ADK and AI
google-adk>=0.1.0
google-generativeai>=0.3.0
google-ai-generativelanguage>=0.5.0

# Google Cloud and Vertex AI
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# Google Auth and API Client
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Data Validation
pydantic>=2.5.0

# Document Processing
PyPDF2>=3.0.0
pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0

# Standard utilities
python-dateutil>=2.8.2

# Optional but recommended for better performance
grpcio>=1.59.0
grpcio-status>=1.59.0

# For local testing (optional)
flask>=3.0.0  # Only needed if using local web interface